#277 and https://github.com/pendulum-chain/spacewalk/pull/486#discussion_r1499498740, https://github.com/pendulum-chain/spacewalk/pull/486#discussion_r1499498740

+++ clients/wallet/src/types.rs
@@ -3,7 +3,7 @@ use primitives::stellar::TransactionEnvelope;
 use std::{collections::HashMap, fmt, fmt::Formatter};
 
 pub type PagingToken = u128;
-pub type Slot = u32;
+pub type Slot = u64;
 pub type StatusCode = u16;
 pub type LedgerTxEnvMap = HashMap<Slot, TransactionEnvelope>;
